ai/requirements.txt
streamlit
numpy
scikit-learn
pillow
webcolors
opencv-python
rembg
colorthief
backgroundremover
requests
